ContactMaps sync with each-other, plus refactoring (see full message)
To cut down on complexity in the higher-level code, the ContactMapCanvas has been re-worked so that
it both sends and recieves selections in the same format that the GeometryCanvas does (i.e. an
array of segment IDs), cleaning up code in places like ViewerPanel and compare.js, with ContactMapCanvas
taking on more complexity. Of course, this format for selections is extremely inefficient for the
Contact Map (although even with 800 segments, there isn't any noticable slowdown).

For now, getting both widgets to use the same format is the first step towards replacing that
format with an actual selection type/class, which would be able to represent the selection in a
much more efficient and easy way.