Contact Map widgets set selection in control panel
Click-and-dragging a selection in one of the Contact Maps will now set the 'Segments' value in the control panel and update the other control panel values accordingly. I tried my best to implement this without further spaghetti-ifying the code, and took the opportunity to do some clean up while I was at it.

You can now call the ControlPanel's 'updateSelection' method from other places in the code, i.e. it's not just called by the event handlers on the text boxes. When calling it, you can include a sourceEvent which will be passed along with any events that it triggers ('segmentChanged', 'locationChanged', 'geneChanged'). You can use the sourceEvent field to filter out certain events and avoid getting stuck in a loop when something calls updateSelection.

While making these changes, I found a few instances of code duplication in regards to various algorithms for converting between ranges and lists of values. I refactored out as many of these as I came across and put them in another module 'Util' for now.