Refactor Selection Management
The way selections of the Genome are handled has been refactored.
There is now a 'Selection' class: An immutable type that represents one
or more selected regions. It can be created from either of three units
we use (locations, segments or genes) and can provide its value in
any of those units. A new 'Controller' class emits events to track
changes in a selection. Different components can listen to events
from the same Controller to keep their selections in sync.