Controller class handles settings along with selection
In preparation for upcoming changes regarding saving/restoring application state,
I have expanded the purpose of the Controller class. Previously, it only handled
syncing the Selection between widgets, but now it handles all application settings
in more-or-less the same way that it handled selections. Since this introduced even
more commonality between widgets, I finally created a new class, `Component`, which
acts as the base calss for all of them (currently, ControlPanel, ContactMapCanvas
and GeometryCanvas). This commit ports those widgets over to this new system and
rewires several of the ControlPanel's settings to use it.

In addition to allowing us to clean even more code out of the main script (compare.js),
this commit provides us with a central location for application state (in the Controller
class), which can help make a lot of future changes more easily.

There is still some work to be done, mainly:

    - The camera position for GeometryCanvas should work through the controller

    - TrackPanel should be made into a Component, and we should add settings for
      the defined tracks.

    - It should be possible to serialize and deserialize all settings/state.

A few other minor changes in this commit worth mentioning:

    - Removed the `AppState` class. This appears to have been an older attempt at
      doing what we are hoping to do next, but isn't actually used anywhere or does anything.

    - Added 'events' to our dependencies. We've had some classes using this for a while now,
      but we must've had access to it as a dependency of something else. Now it's
      officially declared as a dependency in package.json

    - Some fairly small data that is constant for a particular project is now fetched
      when the Project is instantiated, instead of every time we need it. We'll probably
      add more to this over time.