Implement automatic saving/restoring of state and settings
This commit adds serialize and deserialize methods to the Controller class,
which give us the ability to save and restore both the selection and
display settings. This is put into action by the main script (compare.js)
which will save the settings into the browser's local storage and restore
it on page load. The result is seamless preservation of your selection
and settings.

This works well, but there is some more clean-up I would like to do,
as well as implementing settings for the Track Panel before I am
ready to submit a PR. 🙂