Clicking on segments in the GeometryCanvas selects them
The purpose of this is to test out raycasting in preperation for allowing
the user to change the center of rotation. The ultimate goal isn't to set
the selection by clicking on segments, but it serves as a very nice
diagnostic to make sure raycasting is working as it should.

However, clicking on segments actually feels really good! My muscle-memory
makes me want to hold Ctrl or Shift to select multiple segments or ranges
of segments. Something to consider for the future 😁