Clicking on the structure will set the center-of-rotation
Clicking on a segment in the Geometry Canvas will now set the
camera's center of rotation to the point you clicked. To reset,
a "Reset Camera" button has been added to the ControlPanel.

The old functionality of selecting the segment is commented-out
if we ever want it back.

This commit also fixes some funky behavior when a project is first
loaded (and there's no saved state yet). Now everything will be
initialized with the default settings of the Controller class.

I'm really not happy with all the repeated similar code for handling
Controller events. Although I like the API that it gives Components.
A more general solution e.g. `controller.setSetting('variable', x)`
would make it easier to add new settings, but I wouldn't want to
do that unless we could use TypeScript or something to make the
API safer.

It's probably a moot point, since we're planning to rewrite
everything after this next milestone anyway ¯\_(ツ)_/¯