Create Docker Container for building/serving a project
This is the first crack at creating an easy-to-use "app" that can build
and run a browser instance just from project input data. It's not
fool-proof yet, but it's a strong foundation.

To run it:
    - First build the docker image:
        docker build -t 4dgbrunner .
    - Then run with docker-compose
        docker-compose run --rm -p 127.0.0.1:8000:8000 browser

This will "import" the test.01 project, and spin up an instance of the
browser on port 8000.

The important thing here is that, if this Docker image is built
and (eventually) published on DockerHub, users won't even need
this repository to browser their projects. They would only need
to assemble their project in a directory, then run docker-compose
(which can be wrapped in a simple script later).

You can try a bit of this now by moving to a totally different
directory after building the image, copying the 'projects/'
directory into your new directory, then running the docker-compose
command.