Implement "production" version of server container, several QoL changes
There is now a "production" of version of the server container which enables HTTPS
support and implements HTTP Basic Auth. There is a guide in doc/deployment.md which
describes how to setup and configure this version.

Inside the container, Gunicorn is proxied by nginx. Different nginx configs are used
depending on if the image is built for "production" mode or "local" (the default)
mode. Running the server in local mode is now eaiser thanks to the shell script
in bin/run_project. This only requires that the user provide the path to the
directory with their project, and takes care of launching the container
with appropriate arguments. Run 'bin/run_project --help' for more details and
additional arguments.

The Docker image can now also "cache" more than one project (identifed by a project
"name", which is usually just the name of the project directory, but can also be
overridden in the run_project script). This means that multiple projects can be
imported and switched between without the container needing to re-import the project
on startup.