[v1.4.0] Adjustable scaling for contact maps
This commit adds a slider to the control panel that allows
the user to set a "threshold" for the maximum value displayed
in the contact maps. This effectively replaces the 'magnify' option
that was previously set in the project.json

I also switched back to the classic red-and-white color scheme
for the contact maps and fixed the scalar bar so that the minimum
and maximum values will change as the user adjusts the threshold.

In the project.json spec, these changes have obsoleted the 'magnify'
and 'colormap' settings for the contact maps, so they are no longer
used.

As a little bonus, there is also now a "clear settings" button in
the control panel, which will delete the settings saved in Local
Storage then refresh the page.