Add simple pages to view intermediate data
Adds the pages `input_set.html` and `output_set.html` which load and display the
contact sets used as input/output to LAMMPS in the workflow, along with links in
the main browser page to these pages.

Worth noting is that I also added a route to `gtkserver.py` to load arbitary
paths out of the project directory (any path starting with `/project`), so
that I could get at the .tsv files with the data without going through the
whole data-processing rigamaroll.