db_pop now accepts GTK_VERSION environment variable
The hack-y approach to accepting a path to the server directory
as an optional argument (to get the version number) has been
replaced with the much-less hack-y approach to using a
environment variable.

When packaged with nix, db_pop is replaced with a wrapper
script that will have this variable set automatically.

Oh yeah, the script also looks for the version file relative
to its own location as opposed to the current directory lol