Revert "db_pop now accepts GTK_VERSION environment variable"
This reverts commit 4afe9940aac7a56e4a92cd86e6ed41a3e028ee65 (with
some additonal tweaks).

I changed my mind again and decided that, better than being able
to configure the location/value of the version file, it's best
if db_pop just reads it from a known location relative to its
own script (rather than the working directory). The nix package
for db_pop now adds the version file in the expected location
as part of the derivation.